#1093 Added the closest approach dynamo generation for the asteroids pipeline.

+++ asteroid_pipeline.sh
@@ -6,50 +6,90 @@ set -eo pipefail
 # Get the base folder of the system.
 BASE=$(cd "$(dirname "$0")/../.."; pwd)
 
-# Launch the pipeline.
+# Launch the pipeline. It will also create closest_approaches.txt and newly_generated_spks.txt file for creating dynamos.
 cd $BASE/pipelines/asteroids-pipeline
 python3 sync.py -v -o $BASE/sources/asteroids -os $BASE/sources/spice
 
-# Sync the folder to AWS.
-$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-dev/assets/dynamic/asteroids $BASE/sources/asteroids quiet
-$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-staging/assets/dynamic/asteroids $BASE/sources/asteroids quiet
-$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-production/assets/dynamic/asteroids $BASE/sources/asteroids quiet
-$BASE/pipelines/aws-s3-sync/invalidate.py E3JMG193HISS1S "/assets/dynamic/asteroids/*"
-
-sort -u -o $BASE/sources/asteroids/newly_generated_spks.txt $BASE/sources/asteroids/newly_generated_spks.txt
-
-# For every new spk, generate dynamo.
+# For every new closest approach in the closest_approaches file, generate dynamo.
 while read F; do
         IFS=',' read -ra ITEMS <<< "$F"
 		# Get the item info.
 		PIONEER_NAME=${ITEMS[0]}
 		NAIF_ID=${ITEMS[1]}
-		echo "Generating dynamo for $PIONEER_NAME..."
+		START_TIME=${ITEMS[2]}
+		END_TIME=${ITEMS[3]}
+		echo "Generating closest approach dynamo for $PIONEER_NAME..."
 		# Create the dynamo config.
-		cat > $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json <<- EOM
+		cat > $BASE/sources/asteroids/$PIONEER_NAME.earth.orb.json <<- EOM
 		{
 			"type" : "orb",
 			"body" : $NAIF_ID,
-			"otherBody" : 10,
+			"otherBody" : 399,
 			"spiceDirs" : [
 				"lsk",
 				"celestial",
 				"$PIONEER_NAME"
 			],
-			"limitedCoverage" : [null, null]
+			"limitedCoverage" : [$START_TIME, $END_TIME]
 		}
 		EOM
 		# Run dynamo.
-		$BASE/pipelines/dynamogen/dynamogen --spice $BASE/sources/spice --output $BASE/sources/dynamo --config $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json
-		rm $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json
+		$BASE/pipelines/dynamogen/dynamogen --spice $BASE/sources/spice --output $BASE/sources/dynamo --config $BASE/sources/asteroids/$PIONEER_NAME.earth.orb.json
+		rm $BASE/sources/asteroids/$PIONEER_NAME.earth.orb.json
 		# Sync the dynamo folder to AWS.
 		echo "  Uploading to dev, staging, and production AWS..."
 		$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-dev/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
 		$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-staging/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
 		$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-production/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
 		$BASE/pipelines/aws-s3-sync/invalidate.py E3JMG193HISS1S "/assets/dynamic/dynamo/$PIONEER_NAME/*"
-done < $BASE/sources/asteroids/newly_generated_spks.txt
+done < $BASE/sources/asteroids/closest_approaches.txt
 
-# Remove the newly_generated_spks list, since it is appended every time.
-rm $BASE/sources/asteroids/newly_generated_spks.txt
+# Delete the no longer needed closest_approaches file.
+rm -f $BASE/sources/asteroids/closest_approaches.txt
+
+if [[ -f "$BASE/sources/asteroids/newly_generated_spks.txt" ]]; then
+	# Sort and make unique the newly_generated_spks file since it may have picked up multiple runs before being used and removed.
+	sort -u -o $BASE/sources/asteroids/newly_generated_spks.txt $BASE/sources/asteroids/newly_generated_spks.txt
+
+	# For every new spk in the newly_generated_spks file, generate dynamo.
+	while read F; do
+	        IFS=',' read -ra ITEMS <<< "$F"
+			# Get the item info.
+			PIONEER_NAME=${ITEMS[0]}
+			NAIF_ID=${ITEMS[1]}
+			echo "Generating dynamo for $PIONEER_NAME..."
+			# Create the dynamo config.
+			cat > $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json <<- EOM
+			{
+				"type" : "orb",
+				"body" : $NAIF_ID,
+				"otherBody" : 10,
+				"spiceDirs" : [
+					"lsk",
+					"celestial",
+					"$PIONEER_NAME"
+				],
+				"limitedCoverage" : [null, null]
+			}
+			EOM
+			# Run dynamo.
+			$BASE/pipelines/dynamogen/dynamogen --spice $BASE/sources/spice --output $BASE/sources/dynamo --config $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json
+			rm $BASE/sources/asteroids/$PIONEER_NAME.sun.orb.json
+			# Sync the dynamo folder to AWS.
+			echo "  Uploading to dev, staging, and production AWS..."
+			$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-dev/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
+			$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-staging/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
+			$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-production/assets/dynamic/dynamo/$PIONEER_NAME $BASE/sources/dynamo/$PIONEER_NAME quiet
+			$BASE/pipelines/aws-s3-sync/invalidate.py E3JMG193HISS1S "/assets/dynamic/dynamo/$PIONEER_NAME/*"
+	done < $BASE/sources/asteroids/newly_generated_spks.txt
+fi
+
+# Delete the no longer needed newly_generated_spks file.
+rm -f $BASE/sources/asteroids/newly_generated_spks.txt
+
+# Sync the asteroids folder to AWS.
+$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-dev/assets/dynamic/asteroids $BASE/sources/asteroids quiet
+$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-staging/assets/dynamic/asteroids $BASE/sources/asteroids quiet
+$BASE/pipelines/aws-s3-sync/sync.py sync-s3-folder eyes-production/assets/dynamic/asteroids $BASE/sources/asteroids quiet
+$BASE/pipelines/aws-s3-sync/invalidate.py E3JMG193HISS1S "/assets/dynamic/asteroids/*"
